refactor(project): move template tag extraction
