+++ crates/djls-project/src/lib.rs
@@ -7,7 +7,6 @@ mod project;
 mod python;
 mod resolve;
 mod settings;
-mod specs;
 mod sync;
 mod system;
 mod templates;
@@ -44,25 +43,25 @@ pub use settings::django_settings;
 pub use settings::extract_settings;
 pub use settings::settings_module_file;
 pub use settings::template_dirs;
-pub use specs::ArgumentCountConstraint;
-pub use specs::AsVar;
-pub use specs::BlockSpec;
-pub use specs::BlockSpecs;
-pub use specs::ChoiceAt;
-pub use specs::ExtractedDiagnosticConstraint;
-pub use specs::ExtractedDiagnosticMessage;
-pub use specs::ExtractedMessageArg;
-pub use specs::ExtractedMessageTemplate;
-pub use specs::KnownOptions;
-pub use specs::RequiredKeyword;
-pub use specs::SplitPosition;
-pub use specs::TagArgument;
-pub use specs::TagArgumentKind;
-pub use specs::TagRule;
-pub use specs::TagRuleMap;
-pub use specs::extract_block_specs;
-pub use specs::extract_filter_arities;
-pub use specs::extract_tag_rules;
+pub use templates::ArgumentCountConstraint;
+pub use templates::AsVar;
+pub use templates::BlockSpec;
+pub use templates::BlockSpecs;
+pub use templates::ChoiceAt;
+pub use templates::ExtractedDiagnosticConstraint;
+pub use templates::ExtractedDiagnosticMessage;
+pub use templates::ExtractedMessageArg;
+pub use templates::ExtractedMessageTemplate;
+pub use templates::KnownOptions;
+pub use templates::RequiredKeyword;
+pub use templates::SplitPosition;
+pub use templates::TagArgument;
+pub use templates::TagArgumentKind;
+pub use templates::TagRule;
+pub use templates::TagRuleMap;
+pub use templates::extract_block_specs;
+pub use templates::extract_filter_arities;
+pub use templates::extract_tag_rules;
 pub use sync::RefreshData;
 pub use sync::RefreshQuery;
 pub use sync::RefreshQueryResult;
