+++ crates/djls-project/src/templates.rs
@@ -4,9 +4,11 @@ mod libraries;
 mod origins;
 mod registrations;
 mod symbols;
+mod tags;
 
 pub use filters::FilterArity;
 pub use filters::FilterArityMap;
+pub use filters::extract_filter_arities;
 pub use inactive::InactiveLibraries;
 pub use inactive::InactiveLibrary;
 pub use inactive::inactive_template_libraries;
@@ -31,7 +33,26 @@ pub use registrations::RegistrationInfo;
 pub use registrations::RegistrationKind;
 pub(crate) use registrations::TemplateLibraryAnalysis;
 pub use registrations::collect_registrations_from_body;
+pub(crate) use registrations::for_each_registration;
 pub use symbols::SymbolDefinition;
 pub use symbols::SymbolKey;
 pub use symbols::TemplateSymbol;
 pub use symbols::TemplateSymbolKind;
+pub use tags::ArgumentCountConstraint;
+pub use tags::AsVar;
+pub use tags::BlockSpec;
+pub use tags::BlockSpecs;
+pub use tags::ChoiceAt;
+pub use tags::ExtractedDiagnosticConstraint;
+pub use tags::ExtractedDiagnosticMessage;
+pub use tags::ExtractedMessageArg;
+pub use tags::ExtractedMessageTemplate;
+pub use tags::KnownOptions;
+pub use tags::RequiredKeyword;
+pub use tags::SplitPosition;
+pub use tags::TagArgument;
+pub use tags::TagArgumentKind;
+pub use tags::TagRule;
+pub use tags::TagRuleMap;
+pub use tags::extract_block_specs;
+pub use tags::extract_tag_rules;
