+++ crates/djls-project/src/templates/tags/testdata/django_admin_urls.py
