+++ crates/djls-project/src/templates/tags/testdata/django_i18n.py
